backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
chromadb>=1.0.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.1.0
pageindex>=0.1.0
ollama>=0.4.0
openai==1.101.0
pymupdf==1.26.4
tiktoken==0.11.0
pyyaml==6.0.2
